Update from upstream repo aspose-words/Aspose.Words-for-Java

+++ Examples/src/main/java/com/aspose/words/examples/loading_saving/Word2Help.java
@@ -1,10 +1,3 @@
-/* 
- * Copyright 2001-2015 Aspose Pty Ltd. All Rights Reserved.
- *
- * This file is part of Aspose.Words. The source code in this file
- * is only intended as a supplement to the documentation, and is provided
- * "as is", without warranty of any kind, either expressed or implied.
- */
 package com.aspose.words.examples.loading_saving;
 
 import com.aspose.words.*;
@@ -649,7 +642,7 @@ class TopicCollection
             {
                 // Pull up headings. For example: if Heading 1-4 become topics, then I want Headings 5+
                 // to become Headings 4+. Maybe I want to pull up even higher?
-                para.getParagraphFormat().setStyleIdentifier((/*StyleIdentifier*/int)((int)style - 1));
+                para.getParagraphFormat().setStyleIdentifier(style - 1);
             }
         }
